Only install kasmvnc if necessary

+++ kasmvnc/start-template.sh
@@ -7,14 +7,15 @@ if [ -z "${service_port}" ]; then
     displayErrorMessage "ERROR: No service port found in the range \${minPort}-\${maxPort} -- exiting session"
 fi
 
-# FIXME: Only run if kasmvnc is not installed!
-wget https://github.com/kasmtech/KasmVNC/releases/download/v1.3.2/kasmvncserver_oracle_8_1.3.2_x86_64.rpm
-sudo dnf localinstall ./kasmvncserver_*.rpm --allowerasing -y 
-rm ./kasmvncserver_*.rpm
-expect -c 'spawn vncpasswd -u '"${USER}"' -w -r; expect "Password:"; send "password\r"; expect "Verify:"; send "password\r"; expect eof'
-sudo usermod -a -G kasmvnc-cert $USER
-sudo chown $USER /etc/pki/tls/private/kasmvnc.pem
-
+if ! [ -f ${/etc/pki/tls/private/kasmvnc.pem} ]; then
+    # FIXME: Only run if kasmvnc is not installed!
+    wget https://github.com/kasmtech/KasmVNC/releases/download/v1.3.2/kasmvncserver_oracle_8_1.3.2_x86_64.rpm
+    sudo dnf localinstall ./kasmvncserver_*.rpm --allowerasing -y 
+    rm ./kasmvncserver_*.rpm
+    expect -c 'spawn vncpasswd -u '"${USER}"' -w -r; expect "Password:"; send "password\r"; expect "Verify:"; send "password\r"; expect eof'
+    sudo usermod -a -G kasmvnc-cert $USER
+    sudo chown $USER /etc/pki/tls/private/kasmvnc.pem
+fi
 
 
 # Find an available display port
